+++ src/types/spend_bundle.py
@@ -25,10 +25,10 @@ class SpendBundle(Streamable):
     @classmethod
     def aggregate(cls, spend_bundles) -> "SpendBundle":
         coin_solutions: List[CoinSolution] = []
-        sigs = []
-        for _ in spend_bundles:
-            coin_solutions += _.coin_solutions
-            sigs.append(_.aggregated_signature)
+        sigs: List[G2Element] = []
+        for bundle in spend_bundles:
+            coin_solutions += bundle.coin_solutions
+            sigs.append(bundle.aggregated_signature)
         aggregated_signature = AugSchemeMPL.aggregate(sigs)
         return cls(coin_solutions, aggregated_signature)